imgunpkr [fix] add ttf replace support

+++ ImgUnpacker/main.cpp
@@ -10,6 +10,12 @@
 #include <cxxopts.hpp>
 #include <csv.h>
 
+static const char* ITEM_IDX = "Idx";
+static const char* ITEM_NAME = "Name";
+static const char* ITEM_WIDTH = "Width";
+static const char* ITEM_HEIGHT = "Height";
+static const char* ITEM_TYPE = "Type";
+
 namespace fs = std::filesystem;
 
 struct Opts {
@@ -109,16 +115,11 @@ int UnpackImg(const fs::path& in_path, const fs::path& out_path) {
 
     //unpack eifs
     std::ofstream export_list (out_path / "export_list.csv");
-    export_list << "ZipName"
-                << ","
-                << "ImgName"
-                << ","
-                << "ImgWidth"
-                << ","
-                << "ImgHeight"
-                << ","
-                << "ImgType"
-                << std::endl;
+    export_list << ITEM_IDX << ","
+                << ITEM_NAME << ","
+                << ITEM_TYPE << ","
+                << ITEM_WIDTH << ","
+                << ITEM_HEIGHT << std::endl;
 
     int zip_items = img_sec.GetItemsCount(ImageSection::RT_ZIP);
 
@@ -155,16 +156,26 @@ int UnpackImg(const fs::path& in_path, const fs::path& out_path) {
         EIF::EifConverter::eifToBmpFile(eif, p.string());
 
         auto header_p = reinterpret_cast<EIF::EifBaseHeader*>(eif.data());
-        export_list << i << ".zip"
-            << ","
-            << file_stat.m_filename
-            << ","
-            << header_p->width
-            << ","
+        export_list << i << ","
+            << file_stat.m_filename << ","
+            << ToString((image_type)header_p->type) << ","
+            << header_p->width << ","
             << header_p->height
-            << ","
-            << ToString((image_type)header_p->type)
             << std::endl;
+    }
+
+    fs::path ttf_path = out_path/"ttf";
+    fs::create_directory(ttf_path);
+    int ttf_items = img_sec.GetItemsCount(ImageSection::RT_TTF);
+    for(int i = 0; i < ttf_items; i++) {
+        string ttf_name = std::to_string(i) + ".ttf";
+        std::vector<uint8_t> item_bin;
+        img_sec.GetItemData(ImageSection::RT_TTF, i, item_bin);
+        FTUtils::bufferToFile((ttf_path/ttf_name).string(), (char*)item_bin.data(), item_bin.size());
+        export_list << i << ","
+                    << ttf_name << ","
+                    << "TTF" << ","
+                    << 0 << ","  << 0 << std::endl;
     }
 
     export_list.close();
@@ -207,8 +218,6 @@ int vectorToZip(const fs::path& file_path, std::vector<uint8_t>& data, const cha
 
 int PackImg(const fs::path& config_path, const fs::path& vbf_path, const fs::path& out_path) {
 
-    vector<uint8_t> v;
-
     //open vbf and get image section
     VbfFile vbf;
     vbf.OpenFile(vbf_path.string());
@@ -224,41 +233,42 @@ int PackImg(const fs::path& config_path, const fs::path& vbf_path, const fs::pat
     ImageSection img_sec;
     img_sec.Parse(img_sec_bin);
 
-    io::CSVReader<5> in(config_path);
-    in.read_header(io::ignore_extra_column, "ZipName", "ImgName", "ImgWidth", "ImgHeight", "ImgType");
-    std::string zip_name, eif_name, type;
-    uint32_t a,b;
+    io::CSVReader<3> in(config_path);
+    in.read_header(io::ignore_extra_column, ITEM_IDX, ITEM_NAME, ITEM_TYPE);
+    std::string name, type;
+    uint32_t idx;
 
-    int i=0;
-    while(in.read_row(zip_name, eif_name, a, b, type)){
-        //open eif
-        fs::path eif_path = config_path.parent_path() / "eif" / eif_name;
-        vector<uint8_t> eif_bin;
-        FTUtils::fileToVector(eif_path, eif_bin);
+    while(in.read_row(idx, name,type)) {
 
-        //get header data
-        auto eif_header_p = reinterpret_cast<const EIF::EifBaseHeader*>(eif_bin.data());
+        //open resource file
+        fs::path file_path = config_path.parent_path() / ((type == "TTF") ? "ttf" : "eif") / name;
+        vector<uint8_t> file_bin;
+        FTUtils::fileToVector(file_path, file_bin);
 
-        //zip eif
-        std::string tmp_f = std::tmpnam(nullptr);
-        vectorToZip(tmp_f, eif_bin, eif_name.c_str());
+        if(type == "TTF") {
+            img_sec.ReplaceItem(ImageSection::RT_TTF, idx, file_bin);
+        } else {
+            //get header data
+            auto eif_header_p = reinterpret_cast<const EIF::EifBaseHeader*>(file_bin.data());
 
-        vector<uint8_t> zip_bin;
-        FTUtils::fileToVector(tmp_f, zip_bin);
-        //replace
-        img_sec.ReplaceItem(ImageSection::RT_ZIP, i, zip_bin,
-                eif_header_p->width, eif_header_p->height, eif_header_p->type);
+            //zip eif
+            std::string tmp_f = std::tmpnam(nullptr);
+            vectorToZip(tmp_f, file_bin, name.c_str());
 
-        //cleanup
-        fs::remove(tmp_f);
+            vector<uint8_t> zip_bin;
+            FTUtils::fileToVector(tmp_f, zip_bin);
+            fs::remove(tmp_f);
 
-        ++i;
+            //replace
+            img_sec.ReplaceItem(ImageSection::RT_ZIP, idx, zip_bin,
+                                eif_header_p->width, eif_header_p->height, eif_header_p->type);
+        }
     }
 
-    img_sec.SaveToVector(v);
+    img_sec.SaveToVector(img_sec_bin);
 
     //replace vbf image content
-    vbf.ReplaceSectionRaw(1, v);
+    vbf.ReplaceSectionRaw(1, img_sec_bin);
 
     //pack vbf
     vbf.SaveToFile((out_path/"patched.vbf").string());
